Refactoring xr files
