+++ src/webgl/WebGLStrategyCommonMethod.ts
@@ -9,7 +9,7 @@ import Mesh from '../foundation/geometry/Mesh';
 import {Is as is} from '../foundation/misc/Is';
 import ModuleManager from '../foundation/system/ModuleManager';
 import WebGLResourceRepository from './WebGLResourceRepository';
-import {RnXR} from '../rhodonite-xr';
+import {RnXR} from '../xr/main';
 import Vector4 from '../foundation/math/Vector4';
 
 let lastIsTransparentMode: boolean;
